chore: remove unused files, fix misplacements, and clean up dead code
Remove 7 unused production files (StatsTab, ImportExportTab, useModal,
nameUtils, Combobox, RestoreSummary, schemas/index barrel export) and
3 dead exports from settingsUtils. Move tabTreeTypes and tabTreeUtils
from component directory to canonical locations (src/types/, src/utils/).
Rename misnamed story file, fix stale Storybook reference, consolidate
duplicate useStatistics test (porting 4 unique cases), and remove
obsolete pre-Vitest test runner. Update CLAUDE.md to reflect changes.
