+++ src/components/Core/TabTree/GroupEditRow.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Flex, Box, TextField, Button } from '@radix-ui/themes';
 import { getMessage } from '../../../utils/i18n';
 import { ChromeColorPicker } from './ChromeColorPicker';
-import type { ChromeGroupColor } from './tabTreeTypes';
+import type { ChromeGroupColor } from '../../../types/tabTree';
 
 export interface GroupEditRowProps {
   name: string;
