+++ src/components/Core/TabTree/GroupRowBase.tsx
@@ -2,9 +2,9 @@ import React from 'react';
 import { Flex, Text } from '@radix-ui/themes';
 import { ChevronDown, ChevronRight } from 'lucide-react';
 import { getMessage } from '../../../utils/i18n';
-import { chromeGroupColors } from './tabTreeUtils';
+import { chromeGroupColors } from '../../../utils/tabTreeUtils';
 import { AccessibleHighlight } from '../../UI/AccessibleHighlight/AccessibleHighlight';
-import type { ChromeGroupColor } from './tabTreeTypes';
+import type { ChromeGroupColor } from '../../../types/tabTree';
 
 export interface GroupRowBaseProps {
   /** Chrome tab group color */
