+++ src/components/Core/TabTree/TabTree.stories.tsx
@@ -2,7 +2,7 @@ import React, { useState } from 'react';
 import type { Meta, StoryObj } from '@storybook/react';
 import { Theme } from '@radix-ui/themes';
 import { TabTree } from './TabTree';
-import type { TabTreeData } from './tabTreeTypes';
+import type { TabTreeData } from '../../../types/tabTree';
 import { FEATURE_THEMES } from '../../../utils/themeConstants';
 
 /** Sample data: mix of groups and ungrouped tabs */
