+++ src/components/Core/TabTree/TabTree.tsx
@@ -2,8 +2,8 @@ import React, { useMemo, useCallback } from 'react';
 import TreeView, { INodeRendererProps, NodeId } from 'react-accessible-treeview';
 import { Flex, Text, Button, Box, Checkbox, ScrollArea } from '@radix-ui/themes';
 import { getMessage } from '../../../utils/i18n';
-import { TabTreeProps, TabItem, ChromeGroupColor } from './tabTreeTypes';
-import { buildTreeViewData, extractDomain, countTotalTabs } from './tabTreeUtils';
+import { TabTreeProps, TabItem, ChromeGroupColor } from '../../../types/tabTree';
+import { buildTreeViewData, extractDomain, countTotalTabs } from '../../../utils/tabTreeUtils';
 import { TabRowBase } from './TabRowBase';
 import { GroupRowBase } from './GroupRowBase';
