+++ src/components/UI/PopupProfilesList/PopupProfilesList.tsx
@@ -6,7 +6,7 @@ import { getMessage } from '../../../utils/i18n';
 import { loadSessions } from '../../../utils/sessionStorage';
 import { restoreTabs } from '../../../utils/tabRestore';
 import { getRuleCategory } from '../../../schemas/enums';
-import { chromeGroupColors } from '../../Core/TabTree/tabTreeUtils';
+import { chromeGroupColors } from '../../../utils/tabTreeUtils';
 import type { Session } from '../../../types/session';
 
 function getCategoryIcon(categoryId: string | null | undefined): React.ReactNode {
