+++ src/components/UI/SessionWizards/ConflictResolutionStep.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Flex, Text, Box, Badge, RadioGroup, ScrollArea, Callout, Select, Separator } from '@radix-ui/themes';
 import { AlertTriangle } from 'lucide-react';
 import { getMessage, getPluralMessage } from '../../../utils/i18n';
-import { chromeGroupColors } from '../../Core/TabTree/tabTreeUtils';
+import { chromeGroupColors } from '../../../utils/tabTreeUtils';
 import type {
   ConflictAnalysis,
   DuplicateTabAction,
