--- src/schemas/index.ts
@@ -1,4 +1,0 @@
-// Réexportation de tous les schémas et types
-export * from './common.js';
-export * from './enums.js';
-export * from './domainRule.js';
