+++ src/utils/tabCapture.ts
@@ -1,7 +1,7 @@
 import { browser } from 'wxt/browser';
 import { generateUUID } from './utils';
 import type { SavedTab, SavedTabGroup } from '../types/session';
-import type { TabTreeData, TabItem, TabGroupItem, ChromeGroupColor } from '../components/Core/TabTree/tabTreeTypes';
+import type { TabTreeData, TabItem, TabGroupItem, ChromeGroupColor } from '../types/tabTree';
 
 const SYSTEM_URL_PREFIXES = ['chrome://', 'chrome-extension://', 'moz-extension://', 'about:', 'edge://'];
