+++ tests/components/TabTree.test.tsx
@@ -2,7 +2,7 @@ import { describe, it, expect, vi, beforeEach } from 'vitest';
 import { render, screen, fireEvent, within } from '@testing-library/react';
 import { Theme } from '@radix-ui/themes';
 import { TabTree } from '../../src/components/Core/TabTree/TabTree';
-import type { TabTreeData } from '../../src/components/Core/TabTree/tabTreeTypes';
+import type { TabTreeData } from '../../src/types/tabTree';
 
 // Mock i18n
 vi.mock('../../src/utils/i18n', () => ({
